+++ src/cli/invoke.ts
@@ -8,9 +8,6 @@ sourcemapSupport.install({
 	handleUncaughtExceptions: false,
 });
 
-// set up the env
-if (!process.env.NODE_ENV) process.env.NODE_ENV = 'development';
-
 import mri from 'mri';
 
 import type {Args} from './types.js';