+++ src/format.task.ts
@@ -1,4 +1,5 @@
-import {Task, TaskError} from './task/task.js';
+import type {Task} from './task/task.js';
+import {TaskError} from './task/task.js';
 import {printKeyValue} from './utils/print.js';
 import {formatDirectory} from './build/formatDirectory.js';
 import {paths} from './paths.js';