+++ src/task/fixtures/testFailingTaskModule.ts
@@ -1,1 +1,2 @@
 throw Error('for testing purposes');
+export {};